requirements.txt
Flask
flask-cors
flask-login
pillow
watson_developer_cloud